Added path data tokenization
Path data in the d attribute is tokenized as a "meta.instruction-list", if there's a better existing scope name I'd like to hear it.

Each instruction (beginning with a letter) is grouped, and the instruction type, commas, and values are individually tokenized.

I selfishly wanted the letter tokens highlighted automatically, which meant scoping them as a "keyword" which may not be technically correct, but it gets picked up by pretty much every syntax theme so they stand out from the rest of the string.